Support named holes
Add support for named holes of the form `__foo`.

A new phase between `parser` and `namer` looks for all identifiers of
the form `__foo` and replaces them with `???`, annotating the tree with
the name of the hole. Then the main plugin phase extracts the hole name
from the tree annotation when building the log message.