Support holes as function arguments
Given `foo.bar(123, "abc", ???)`, we look up the `foo.bar` symbol to get
the types of its parameters, which gives us the type of the hole.

Note: there's a big assumption here that the number of params exactly
matches the number of args, which is not true in the case of default
args, varargs, etc. So this will not catch all cases, but it's a step in
the right direction.